backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
pydantic==2.10.4
pydantic-settings==2.7.0

# Base de datos
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.21.0

# Redis y caché
redis==5.2.1
celery==5.4.0
flower==2.0.1

# ARCA/AFIP
cryptography==44.0.0
lxml==5.3.0
zeep==4.3.1
requests==2.32.3

# Bank-Kit (PDF)
reportlab==4.2.5

# OCR e IA
google-cloud-aiplatform==1.78.0
Pillow==11.1.0
pytesseract==0.3.13

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
itsdangerous==2.2.0

# Utilidades
python-dateutil==2.9.0
pytz==2024.2
click==8.1.8
httpx==0.28.1
tenacity==9.0.0

# Monitoreo y logging
structlog==24.4.0
python-json-logger==2.0.7
prometheus-client==0.21.1

# Tests
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
faker==33.1.0

# Calidad de código
ruff==0.9.1
mypy==1.14.1
